Melhorando as buscas locais - vnd esta funcionando com os movimentos concertados - falta melhorar a construção da solução: analisar tempos de antecipação das solicitações, qual ficaria melhor e se algum ficaria? - melhorar sequencia de embarque: qual a melhor forma?
- corrigir avaliação do numero de veiculos
- falta olhar a parte de rotas reduzidas, só está fazendo o movimento básico

- CORRIGIR VIZINHANÇA DE REALOCAÇÃO DE SOLICITAÇÕES -> a solicitação não é retirada da primeira rota e é inserida na segunda, assim o valor da função objetivo vai ser sempre maior, fazendo com que não haja nenhuma solução que seja melhorada com essa vizinhança
- CORREÇÃO ANTERIOR ESTÁ SENDO FEITA + falta arrumar as rotas
- implementação do algoritmo VND
- falta implementar o algoritmo de pós-otimização
- falta testar e achar onde esta o erro da repetição de solicitações em rotas distintas

+++ src/main/java/ProblemRepresentation/Request.java
@@ -360,6 +360,6 @@ public class Request implements Cloneable {
     public int hashCode() {
         String string = Integer.toString(this.getId());
         int hash = string.hashCode();
-        return id;
+        return id.intValue();
     }
 }
